apps/legal_discovery/requirements.txt
neuro-san
flask
flask-socketio
python-socketio
schedule
gunicorn
eventlet
Pillow
PyMuPDF
pytesseract
PyPDF2
chromadb
rfc3987
neo4j
requests
pandas
beautifulsoup4
python-docx
python-pptx
flask-sqlalchemy
psycopg2-binary
langchain-google-genai
google-generativeai
pyvis
networkx
spacy
weasyprint
reportlab
pygraphviz
gTTS

en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

